Moved isABITypeStaticSized() to abi.ts module. Small readability tweaks in few other modules.
