+++ src/debug/sol_debugger.ts
@@ -175,8 +175,7 @@ export interface EventDesc {
 /**
  * TODO(dimo): Make memory and storage be computed only for instructions that change them, and for all other
  * instructions alias the previous steps' memory/storage
- */
-/**
+ *
  * Low-level machine state at a given trace step. It directly mirrors the state reported from Web3
  * and doesn't include any higher-level information that requires debug info.
  */
